feat: add Binance crypto trading support (broker, CLI, paper engine)
Extends the platform to support crypto trading via Binance Spot:
- BinanceClient for market data and order execution
- Crypto symbol resolution (BTCUSDT-style pairs)
- Paper engine updates for fractional quantities and crypto pricing
- CLI commands for crypto analysis and trading
- Safety checker and concentration limits adapted for crypto assets

+++ skopaq/db/models.py
@@ -23,7 +23,7 @@ class TradeRecord(BaseModel):
     symbol: str
     exchange: str = "NSE"
     side: str
-    quantity: int
+    quantity: Decimal
     price: Optional[Decimal] = None
     order_type: str = "MARKET"
     product: str = "CNC"
